Fix issue #175: Go core: split CLI orchestration code into smaller files
